+++ changelog.md
@@ -21,6 +21,7 @@ Codion Change Log
 ### is.codion.swing.framework.ui
 - EntityTablePanel, EntityDialogs, minor i18n message improvements.
 - EntityTableExportPanel split from EntityTableExport, related refactoring and minor improvements.
+- EntityTableExportPanel configuration now included in user preferences.
 
 ## 0.18.51
 ### is.codion.swing.common.ui